Restructure resolvers
